refactor(kernel): move ledger materialization
