+++ packages/kernel/src/store/index.ts
@@ -1,7 +1,7 @@
 export * from "../persistence/blob/content-addressed-blob-store.ts";
 export * from "../integrity/stable-hash.ts";
 export * from "./daemon-runtime.ts";
-export * from "./ledger-materializer.ts";
+export * from "../write-coordination/materialization/ledger-materializer.ts";
 export * from "../persistence/local/local-lock-registry.ts";
 export * from "../persistence/git/local-version-control-system.ts";
 export * from "../persistence/markdown/markdown-artifact-store.ts";
